# docs/specs/swagger.json
{"openapi":"3.0.1","info":{"title":"'lciii-scaffolding'","description":"'LCII Initial Scaffolding'","contact":{"name":"None","email":"none@none.com"},"version":"'0.0.1-SNAPSHOT'"},"servers":[{"url":"http://localhost:8080","description":"'LCII Initial Scaffolding'"}],"paths":{"/api/countries":{"get":{"tags":["country-controller"],"operationId":"getPaises","parameters":[{"name":"nombre","in":"query","required":false,"schema":{"type":"string"}},{"name":"codigo","in":"query","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CountryDTO"}}}}}}},"post":{"tags":["country-controller"],"operationId":"postPaisesAleatorios","requestBody":{"content":{"application/json":{"schema":{"type":"integer","format":"int32"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CountryDTO"}}}}}}}},"/api/countries/{language}/language":{"get":{"tags":["country-controller"],"operationId":"getPaisesPorIdioma","parameters":[{"name":"language","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CountryDTO"}}}}}}}},"/api/countries/{continent}/continent":{"get":{"tags":["country-controller"],"operationId":"getPaisesEnContinente","parameters":[{"name":"continent","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CountryDTO"}}}}}}}},"/api/countries/tests":{"get":{"tags":["country-controller"],"operationId":"obtenerPaisesTodos","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Country"}}}}}}}},"/api/countries/most-borders":{"get":{"tags":["country-controller"],"operationId":"getPaisConMasFronteras","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CountryDTO"}}}}}}}},"components":{"schemas":{"CountryDTO":{"type":"object","properties":{"code":{"type":"string"},"name":{"type":"string"}}},"Country":{"type":"object","properties":{"name":{"type":"string"},"population":{"type":"integer","format":"int64"},"area":{"type":"number","format":"double"},"code":{"type":"string"},"region":{"type":"string"},"borders":{"type":"array","items":{"type":"string"}},"languages":{"type":"object","additionalProperties":{"type":"string"}}}}}}}